+++ app/src/main/java/com/fallTurtle/myrestaurantgallery/ui/locationList/LocationListActivity.kt
@@ -10,7 +10,6 @@ import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.fallTurtle.myrestaurantgallery.R
-import com.fallTurtle.myrestaurantgallery.ui.adapter.LocationAdapter
 import com.fallTurtle.myrestaurantgallery.databinding.ActivityLocationListBinding
 import com.fallTurtle.myrestaurantgallery.ui.dialog.ProgressDialog
 import com.fallTurtle.myrestaurantgallery.data.etc.NetworkWatcher
